requirements.txt
mysql-connector-python==8.0.11
Flask==0.12.2
